cleaned code, remove unused imports and added necessary comments
