+++ client/src/components/layout/AuthNavbar.jsx
@@ -1,5 +1,7 @@
 import logo from "../../../public/images/light-logo.svg";
 import { Link } from "react-router-dom";
+// Navbar for authentication
+// if inv = true, inverse the colors of the text and border
 const AuthNavbar = ({ inv }) => {
   return (
     <div className="px-[10px] lg:px-[30px] lg:bg-transparent pt-[10px] py-[40px] font-bold top-0  fixed w-full h-[80px] text-[14px] xl:text-[20px] grid-cols-[4fr,1fr] gap-[10px] lg:gap-0 md:grid-cols-[5fr,1.5fr] lg:grid-cols-[7fr,1fr] grid justify-center items-center">
