+++ client/src/components/layout/Footer.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Footer = () => {
   return (
     <div className="w-screen font-merryweather text-[#E7E4D5] font-bold  h-[14vh] bg-[#395143] flex flex-col justify-end items-center pb-[20px]">
