+++ client/src/components/layout/LobbyNav.jsx
@@ -3,9 +3,11 @@ import axios from "axios";
 import Cookies from "js-cookie";
 import logo from "../../../public/images/dark-logo.svg";
 
+// Navbar for lobbies
 const LobbyNav = () => {
   //get the access token
   const token = Cookies.get("authToken");
+  // function for calling the log out API
   const handleLogOut = async () => {
     try {
       const response = await axios.get("http://127.0.0.1:5000/logout/", {
@@ -23,7 +25,7 @@ const LobbyNav = () => {
     }
   };
   return (
-    <div className="px-[10px] lg:px-[30px] lg:bg-transparent pt-[10px] py-[40px] font-bold top-0  fixed w-full h-[80px] text-[14px] xl:text-[20px] grid-cols-[4fr,1fr] gap-[10px] lg:gap-0 md:grid-cols-[5fr,1.5fr] lg:grid-cols-[6fr,1fr] grid justify-center items-center">
+    <div className="px-[10px] lg:px-[30px] lg:bg-transparent pt-[10px] py-[40px] font-bold top-0  fixed w-full h-[80px] text-[14px] xl:text-[20px] grid-cols-[4fr,2fr] gap-[10px] lg:gap-0 md:grid-cols-[5fr,1.5fr] lg:grid-cols-[6fr,1fr] grid justify-center items-center">
       <div className="w-[150px] lg:w-[200px] flex items-center h-full flex-col justify-center ">
         <Link to="/">
           <img src={logo} />
