+++ client/src/components/layout/Navbar.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import logo from "../../../public/images/light-logo.svg";
-import line from "../../../public/images/main/yellow-line.svg";
+// Outer navbar for landing page
 const Navbar = () => {
   return (
     <div className="px-[10px] lg:px-[30px] py-[10px] z-20 font-bold top-0 bg-[#395143] fixed w-full h-[70px] text-[14px] xl:text-[20px] grid-cols-[1fr,0.5fr] gap-[10px] lg:gap-0 lg:grid-cols-[5fr,1fr,1fr] grid justify-center items-center">
