--- client/src/data/data.js
@@ -1,217 +1,0 @@
-// this file contains dummy data for testing
-
-export const profile = {
-  firstName: "meow",
-  lastName: "meow",
-  pseudo: "meow",
-  email: "meow@gmail.com",
-};
-
-export const moderateurs = [
-  {
-    imgUrl: "https://example.com/avatar1.jpg",
-    name: "John Doe",
-    username: "john_doe",
-    email: "john.doe@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar2.jpg",
-    name: "Jane Smith",
-    username: "jane_smith",
-    email: "jane.smith@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar1.jpg",
-    name: "John Doe",
-    username: "john_doe",
-    email: "john.doe@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar2.jpg",
-    name: "Jane Smith",
-    username: "jane_smith",
-    email: "jane.smith@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar1.jpg",
-    name: "John Doe",
-    username: "john_doe",
-    email: "john.doe@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar2.jpg",
-    name: "Jane Smith",
-    username: "jane_smith",
-    email: "jane.smith@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar1.jpg",
-    name: "John Doe",
-    username: "john_doe",
-    email: "john.doe@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar2.jpg",
-    name: "Jane Smith",
-    username: "jane_smith",
-    email: "jane.smith@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar1.jpg",
-    name: "John Doe",
-    username: "john_doe",
-    email: "john.doe@example.com",
-  },
-  {
-    imgUrl: "https://example.com/avatar2.jpg",
-    name: "Jane Smith",
-    username: "jane_smith",
-    email: "jane.smith@example.com",
-  },
-];
-
-export const articles = [
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-  {
-    title:
-      "Nicotinamide mononucleotide (NMN) as an anti-aging health product – Promises and safety concerns",
-    date: "20/30/2023",
-    resume:
-      "popo on kaki popo kakalant kokokkokokookokokookokefh slifguqzenyfuiqzeybfuiezybdfv zeyfieudsbfiueydfniedsfinuzeo, iflusdjfoinlzeilufniodsfu heiuzbfyoizefsubienfoiezfune oifunozeifuiezfneiofnu zeioffsfihfoimzyfzbyfiusbfiunfiousnff",
-    url: "hahahahhaha.pdf",
-    auteurs: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    institutions: ["boubout mamat", "mimiw", "mimus", "tchat"],
-    motscle: ["mimiw", "mimus", "bchbouchet"],
-    refs: ["ninit", "ninit"],
-    text: "In the fast-evolving landscape of technology, quantum computing stands out as a groundbreaking frontier, promising to revolutionize the way we process information. Unlike classical computers that rely on bits to represent either 0 or 1, quantum computers use quantum bits or qubits, leveraging the principles of quantum mechanics to perform complex computations at unprecedented speeds. Quantum Bits: The Building Blocks of Quantum ComputingAt the heart of quantum computing are qubits, which, thanks to the phenomenon of superposition, can exist in multiple states simultaneously. This unique property allows quantum computers to process a vast number of possibilities in parallel, leading to an exponential increase in computational power compared to classical computers.Entanglement: Quantum Computing's Secret SauceEntanglement is another fundamental quantum principle that sets quantum computing apart. When qubits become entangled, the state of one qubit becomes directly linked to the state of another, regardless of the physical distance between them. This interdependence enables quantum computers to solve certain problems more efficiently than classical computers.Quantum Supremacy: A Milestone Achieved. In recent years, there has been a race among researchers and tech giants to achieve quantum supremacy—the point at which a quantum computer can outperform the most powerful classical computers. Google's announcement of achieving quantum supremacy with its 53-qubit Sycamore processor in 2019 marked a significant milestone, showcasing the potential of quantum computing to solve complex problems beyond the reach of classical machines.Applications Across Industrieslecular interactions could significantly accelerate the identification of new drugs and materials.",
-  },
-];
